+++ src/main/java/org/folio/entlinks/controller/delegate/LinkingServiceDelegate.java
@@ -25,6 +25,7 @@ import org.folio.entlinks.exception.RequestBodyValidationException;
 import org.folio.entlinks.integration.internal.InstanceStorageService;
 import org.folio.entlinks.service.consortium.propagation.ConsortiumLinksPropagationService;
 import org.folio.entlinks.service.consortium.propagation.ConsortiumPropagationService;
+import org.folio.entlinks.service.consortium.propagation.model.LinksPropagationData;
 import org.folio.entlinks.service.links.InstanceAuthorityLinkingService;
 import org.folio.spring.FolioExecutionContext;
 import org.folio.tenant.domain.dto.Parameter;
@@ -73,8 +74,9 @@ public class LinkingServiceDelegate {
     validateLinks(instanceId, links);
     var incomingLinks = mapper.convertDto(links);
     linkingService.updateLinks(instanceId, incomingLinks);
-    propagationService.propagate(incomingLinks, ConsortiumPropagationService.PropagationType.UPDATE,
-      context.getTenantId());
+    var propagationData = new LinksPropagationData(instanceId, incomingLinks);
+    propagationService.propagate(propagationData, ConsortiumPropagationService.PropagationType.UPDATE,
+        context.getTenantId());
   }
 
   public LinksCountDtoCollection countLinksByAuthorityIds(UuidCollection authorityIdCollection) {