+++ src/main/java/org/folio/entlinks/service/consortium/propagation/ConsortiumLinksPropagationService.java
@@ -1,14 +1,13 @@
 package org.folio.entlinks.service.consortium.propagation;
 
-import java.util.List;
-import org.folio.entlinks.domain.entity.InstanceAuthorityLink;
 import org.folio.entlinks.service.consortium.ConsortiumTenantsService;
+import org.folio.entlinks.service.consortium.propagation.model.LinksPropagationData;
 import org.folio.entlinks.service.links.InstanceAuthorityLinkingService;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ConsortiumLinksPropagationService extends ConsortiumPropagationService<List<InstanceAuthorityLink>> {
+public class ConsortiumLinksPropagationService extends ConsortiumPropagationService<LinksPropagationData> {
 
   private static final String ILLEGAL_PROPAGATION_MSG = "Propagation type '%s' is not supported for links.";
 
@@ -22,11 +21,11 @@ public class ConsortiumLinksPropagationService extends ConsortiumPropagationServ
   }
 
   @Override
-  protected void doPropagation(List<InstanceAuthorityLink> links,
+  protected void doPropagation(LinksPropagationData propagationData,
                                PropagationType propagationType) {
     switch (propagationType) {
       case CREATE, DELETE -> throw new IllegalArgumentException(ILLEGAL_PROPAGATION_MSG.formatted(propagationType));
-      case UPDATE -> instanceAuthorityLinkingService.updateLinks(links.get(0).getInstanceId(), links);
+      case UPDATE -> instanceAuthorityLinkingService.updateLinks(propagationData.instanceId(), propagationData.links());
       default -> throw new IllegalStateException("Unexpected value: " + propagationType);
     }
   }