+++ src/test/java/org/folio/entlinks/service/consortium/ConsortiumLinksPropagationServiceTest.java
@@ -1,5 +1,6 @@
 package org.folio.entlinks.service.consortium;
 
+import static java.util.Collections.emptyList;
 import static org.folio.support.base.TestConstants.TENANT_ID;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -16,6 +17,7 @@ import org.folio.entlinks.domain.entity.InstanceAuthorityLink;
 import org.folio.entlinks.exception.FolioIntegrationException;
 import org.folio.entlinks.service.consortium.propagation.ConsortiumAuthorityPropagationService;
 import org.folio.entlinks.service.consortium.propagation.ConsortiumLinksPropagationService;
+import org.folio.entlinks.service.consortium.propagation.model.LinksPropagationData;
 import org.folio.entlinks.service.links.InstanceAuthorityLinkingService;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.folio.spring.test.type.UnitTest;
@@ -40,9 +42,11 @@ class ConsortiumLinksPropagationServiceTest {
     var link = new InstanceAuthorityLink();
     link.setInstanceId(instanceId);
     List<InstanceAuthorityLink> links = List.of(link);
+    final var propagationData = new LinksPropagationData(instanceId, links);
 
     doMocks();
-    propagationService.propagate(links, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
+    propagationService.propagate(
+        propagationData, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
 
     verify(tenantsService).getConsortiumTenants(TENANT_ID);
     verify(executionService, times(3)).executeAsyncSystemUserScoped(any(), any());
@@ -51,12 +55,12 @@ class ConsortiumLinksPropagationServiceTest {
 
   @Test
   void testPropagateIllegalPropagationType() {
-    List<InstanceAuthorityLink> links = List.of(new InstanceAuthorityLink());
+    final var propagationData = new LinksPropagationData(null, emptyList());
 
     doMocks();
 
     var exception = assertThrows(IllegalArgumentException.class,
-      () -> propagationService.propagate(links, ConsortiumAuthorityPropagationService.PropagationType.CREATE,
+      () -> propagationService.propagate(propagationData, ConsortiumAuthorityPropagationService.PropagationType.CREATE,
         TENANT_ID));
 
     assertEquals("Propagation type 'CREATE' is not supported for links.", exception.getMessage());
@@ -66,9 +70,10 @@ class ConsortiumLinksPropagationServiceTest {
   void testPropagateException() {
     doThrow(FolioIntegrationException.class).when(tenantsService).getConsortiumTenants(any());
 
-    List<InstanceAuthorityLink> links = List.of(new InstanceAuthorityLink());
+    final var propagationData = new LinksPropagationData(null, emptyList());
 
-    propagationService.propagate(links, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
+    propagationService.propagate(
+        propagationData, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
 
     verify(tenantsService, times(1)).getConsortiumTenants(any());
     verify(executionService, times(0)).executeAsyncSystemUserScoped(any(), any());